capic: modify to build and run with systemd v221 and later
The initial CAPIC development was done with the systemd release v219.  Since then, a stable version of the sd-bus and sd-event API was released in v221.  The API and ABI changes between these releases include:

* v219 -> v220
  - sd-event.h: SD_EVENT_PASSIVE -> SD_EVENT_INITIAL, SD_EVENT_PREPARED -> SD_EVENT_ARMED
  - sd-bus.h: sd_bus_message_handler_t signature changed
  - sd-bus.h: several functions added, one signature parameter re-named
* v220 -> v221
  - libsystemd: sd_bus_get_scope() exported
  - sd-bus.h: sd_bus_call_method_async() introduced

The shared and to-be generated sources were modified to support the stable API/ABI (as of v221) and the necessary compatibility macros were added.  The shared library capic makes all necessary checks in its configure script and uses the results in its private.h.  The generated code of reference applications relies on capic/dbus-private.h that uses the value of CC_SD_API_VERSION.  The latter is defined as 221 by default and must be overridden during build to support earlier systemd versions.

Signed-off-by: Pavel Konopelko <pkonopel@visteon.com>