tools: extend core.cli to open file and pass it to Generator
Invoking the Generator requires that the core plugin can be accessed.  This
is achieved by extending core.cli.feature to include it.  The bundle
org.eclipse.core.resources (which is required to manipulate Eclipse
workspace and projects) must be included into the feature as well.

In order to pass an instance of IFile to Generator, the corresponding input
file must be linked into the default workspace that is opened upon the
startup of the rcp application.  The corresponding Eclipse functionality
relies on absolute paths.  Therefore the generator binary expects one
command line argument and it must be an existing absolute path.  Any changes
to the default workspace must be reverted.  Otherwise subsequent launches
with the same argument will result into 'overlapping resource' errors
because the file has been already linked.

As the result:

* core.cli plugin launches clean from the ide using launch configuration
  'org.genivi.capic.core.cli.appid'.

* core.product does NOT launch from the ide and the exported binary does not
  launch either due to unresolved org.eclipse.resources; in the latter case
  config.ini does not include org.eclipse.core.resources in its osgi.bundles
  entry; the failure is consistent regardless what target platform is active
  ('org.genivi.capic.core.target' or 'Running Platform'); both target
  platforms include org.eclipse.resources in their content.

* core.product binaries built with maven launch without error on both linux
  and windows.

Signed-off-by: Pavel Konopelko <pkonopel@visteon.com>