tools: introduce IFileMaker as attribute of Generator
The Generator needs a way to create folders and files for its output.  The
absolute path and location for these entities, though, depends on whether
the generator has been invoked from the workbench or from the standalone cli
binary.  The interface IFileMaker allows generator clients to pass their
policy for creating files.

When invoking the Generator from the workbench, the output files are added
to the project that contains the input file.  They are placed into the
folder that is named by default 'src-gen' and is added directly under the
project.

When invoking the Generator from the standalone cli binary, the output files
are created in the subdirectory of the current working directory.  The
subdirectory is named by default 'src-gen'.

In both cases, the contents of already existing files is overwritten.  Any
other content of the output directories is preserved.

Signed-off-by: Pavel Konopelko <pkonopel@visteon.com>