tools: integrate xtend-based generator into core
This requires additional plugin dependencies as well as an implementation of
Franca model loading.  The latter currently relies on a shortcut solution
implemented as a method on core.Generator that cannot properly handle import
statements.  Proper solution must rely on FrancaPersistenceManager class,
which supports loading and saving multi-file Franca models.

Xtend/Xtext plugins rely on org.apache.log4j.  The logger requires
initialization parameters that must be explicitly provided in the case of
the standalone binary.

Signed-off-by: Pavel Konopelko <pkonopel@visteon.com>