tools: modify core.ui and core.cli to use dependency injection
Franca and Xtext plugins rely on the dependency injection for their
initialization.  In order to use it, the top-level objects in the hierarchy
must be created via injectors (as contrasted to the explicit invocation of
their constructors via new).  This ensures that all their dependencies will
be properly injected as well.

* When Franca plugins are used inside the Eclipse IDE, the injector must be
  provided by Franca / Xtext [1].  The client plugin provides an
  implementation of ExecutableExtensionFactory and refers to it in its
  declaration of the extension handler.

* When Franca plugins are used in a standalone application, the injector
  must be created by FrancaIDLStandaloneSetup [2].  The application uses the
  injector to create an instance of Generator.

The modifications described above enables the core.Generator to use the
FrancaPersistenceManager functionality.  The latter equips the model loader
with the capability to resolve import statements in Franca, which is now
supported by the Common API C code generator.

[1] https://groups.google.com/d/msg/franca-framework/TntSXYuLiVY/xvEM8WLZCAAJ
[2] https://groups.google.com/d/msg/franca-framework/5IjfJJll4j8/AQHfL_-TCAAJ

Signed-off-by: Pavel Konopelko <pkonopel@visteon.com>