tools: introduce deploy script for publishing binary artifacts
The purpose of this script is to package the build outputs and arrange them
such that they can be directly hosted on GENIVI docs web [1] and provide
multiple release versions simultaneously.  Standalone code generators for
various platforms are re-packaged into one archive and arranged in a
directory structure organized by the release version.  Update site for
Eclipse plug-ins is structured as a p2 composite repository that also
includes multiple releases.

[1] http://docs.projects.genivi.org/

Signed-off-by: Pavel Konopelko <pkonopel@visteon.com>