perf: introduce capic performance benchmark
Measure the maximum throughput when synchronously sending messages with
and without parameters from client to server.  Synchronously means that
the client only sends the next message after it has received the
response for the previous one.

The build script relies on `capic-core-gen` to generate code from .fidl
file.  The corresponding binary must be accessible through PATH during
the build.

Signed-off-by: Pavel Konopelko <pkonopel@visteon.com>