all: fix build scripts to avoid warnings with gcc <5.0
When using sd-bus with gcc <5.0, the macro SD_BUS_METHOD triggers the
warning -Wmissing-field-initializers although the code looks correct.
Since this warning is included in -Wextra, the build fails due to the
-Werror option.  The warning seems to be caused by a bug in gcc [1] that
was fixed starting with 5.0.  For the gcc versions prior to 5.0, this
warning is disabled as a work around.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=60784

Signed-off-by: Pavel Konopelko <pkonopel@visteon.com>